RawTech exporter upgraded
The exporter can now handle external techs exporting and loading.  Player Techs can also spawn as enemies if they are aircraft or are anchored bases.
Check the Raw Techs folder on export (there should be a button to export the player Tech when paused)

Mostly intended for use with the upcoming Sub_Missions system while Population Injector is being worked on